Conflicts:
	tests/test_apps/genqa/run.sh

+++ src/frontend/org/voltdb/expressions/HashRangeExpression.java
@@ -40,7 +40,7 @@ public class HashRangeExpression extends AbstractValueExpression {
         RANGE_END
     }
 
-    protected ImmutableSortedMap<Long, Long> m_ranges;
+    protected ImmutableSortedMap<Integer, Integer> m_ranges;
     protected int m_hashColumn = Integer.MIN_VALUE;
 
     public HashRangeExpression() {
@@ -89,14 +89,14 @@ public class HashRangeExpression extends AbstractValueExpression {
     /**
      * @return the ranges
      */
-    public Map<Long, Long> getRanges() {
+    public Map<Integer, Integer> getRanges() {
         return m_ranges;
     }
 
     /**
      * @param ranges the column_alias to set
      */
-    public void setRanges(Map<Long, Long> ranges) {
+    public void setRanges(Map<Integer, Integer> ranges) {
         m_ranges = ImmutableSortedMap.copyOf(ranges);
     }
 
@@ -145,10 +145,10 @@ public class HashRangeExpression extends AbstractValueExpression {
         super.toJSONString(stringer);
         stringer.key(Members.HASH_COLUMN.name()).value(m_hashColumn);
         stringer.key(Members.RANGES.name()).array();
-        for (Map.Entry<Long, Long> e : m_ranges.entrySet()) {
+        for (Map.Entry<Integer, Integer> e : m_ranges.entrySet()) {
             stringer.object();
-            stringer.key(Members.RANGE_START.name()).value(e.getKey());
-            stringer.key(Members.RANGE_END.name()).value(e.getValue());
+            stringer.key(Members.RANGE_START.name()).value(e.getKey().intValue());
+            stringer.key(Members.RANGE_END.name()).value(e.getValue().intValue());
             stringer.endObject();
         }
         stringer.endArray();
@@ -158,10 +158,10 @@ public class HashRangeExpression extends AbstractValueExpression {
     protected void loadFromJSONObject(JSONObject obj) throws JSONException {
         m_hashColumn = obj.getInt(Members.HASH_COLUMN.name());
         JSONArray array = obj.getJSONArray(Members.RANGES.name());
-        ImmutableSortedMap.Builder<Long, Long> b = ImmutableSortedMap.naturalOrder();
+        ImmutableSortedMap.Builder<Integer, Integer> b = ImmutableSortedMap.naturalOrder();
         for (int ii = 0; ii < array.length(); ii++) {
             JSONObject range = array.getJSONObject(ii);
-            b.put(range.getLong(Members.RANGE_START.name()), range.getLong(Members.RANGE_END.name()));
+            b.put(range.getInt(Members.RANGE_START.name()), range.getInt(Members.RANGE_END.name()));
         }
         m_ranges = b.build();
     }
